some refactoring for Logout - the logout function moved to Utilities.js - the logout function for the tab close event moved to Utilities.js as well
